feat: add curve type, reorganize types
