+++ packages/crypto/src/index.ts
@@ -1,5 +1,5 @@
-export { KeyType } from './constants';
-export { KeyPair, KeyPairString } from './key_pair';
+export { CurveType, KeyPairString, KeyType } from './constants';
+export { KeyPair } from './key_pair';
 export { Signature } from './key_pair_base';
 export { KeyPairEd25519 } from './key_pair_ed25519';
 export { KeyPairSecp256k1 } from './key_pair_secp256k1';
